+++ tests/PhpWord/Writer/Word2007/ElementTest.php
@@ -19,8 +19,8 @@ namespace PhpOffice\PhpWord\Writer\Word2007;
 
 use PhpOffice\Common\XMLWriter;
 use PhpOffice\PhpWord\Element\Comment;
-use PhpOffice\PhpWord\Element\Text;
 use PhpOffice\PhpWord\Element\TextRun;
+use PhpOffice\PhpWord\Element\TrackChange;
 use PhpOffice\PhpWord\PhpWord;
 use PhpOffice\PhpWord\TestHelperDOCX;
 
@@ -414,5 +414,21 @@ class ElementTest extends \PHPUnit\Framework\TestCase
         $this->assertTrue($doc->elementExists('/w:document/w:body/w:p/w:commentRangeStart'));
         $this->assertTrue($doc->elementExists('/w:document/w:body/w:p/w:commentRangeEnd'));
         $this->assertTrue($doc->elementExists('/w:document/w:body/w:p/w:r/w:commentReference'));
+    }
+
+    public function testTrackChange()
+    {
+        $phpWord = new PhpWord();
+        $section = $phpWord->addSection();
+        $text = $section->addText('my dummy text');
+        $text->setChangeInfo(TrackChange::INSERTED, 'author name');
+        $text2 = $section->addText('my other text');
+        $text2->setTrackChange(new TrackChange(TrackChange::DELETED, 'another author', new \DateTime()));
+
+        $doc = TestHelperDOCX::getDocument($phpWord);
+
+        $this->assertTrue($doc->elementExists('/w:document/w:body/w:p/w:ins/w:r'));
+        $this->assertEquals('author name', $doc->getElementAttribute('/w:document/w:body/w:p/w:ins', 'w:author'));
+        $this->assertTrue($doc->elementExists('/w:document/w:body/w:p/w:del/w:r/w:delText'));
     }
 }