+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v0.15.0 (?? ??? 2018)
 ### Added
 - Parsing of "align" HTML attribute - @troosan #1231
 - Parse formatting inside HTML lists - @troosan @samimussbach #1239 #945 #1215 #508
+- Add support for Track changes @Cip @troosan #354 #1262
 
 ### Fixed
 - fix reading of docx default style - @troosan #1238